+++ ios/Plugin/IntercomPlugin.m
@@ -3,7 +3,7 @@
 
 // Define the plugin using the CAP_PLUGIN Macro, and
 // each method the plugin supports using the CAP_PLUGIN_METHOD macro.
-CAP_PLUGIN(IntercomPlugin, "IntercomPlugin",
+CAP_PLUGIN(IntercomPlugin, "Intercom",
            CAP_PLUGIN_METHOD(registerIdentifiedUser, CAPPluginReturnPromise);
            CAP_PLUGIN_METHOD(registerUnidentifiedUser, CAPPluginReturnPromise);
            CAP_PLUGIN_METHOD(updateUser, CAPPluginReturnPromise);
@@ -17,6 +17,7 @@ CAP_PLUGIN(IntercomPlugin, "IntercomPlugin",
            CAP_PLUGIN_METHOD(hideLauncher, CAPPluginReturnPromise);
            CAP_PLUGIN_METHOD(displayInAppMessages, CAPPluginReturnPromise);
            CAP_PLUGIN_METHOD(hideInAppMessages, CAPPluginReturnPromise);
+           CAP_PLUGIN_METHOD(displayCarousel, CAPPluginReturnPromise);
            CAP_PLUGIN_METHOD(setUserHash, CAPPluginReturnPromise);
            CAP_PLUGIN_METHOD(setBottomPadding, CAPPluginReturnPromise);
 )